requirements.txt
Flask
requests
psycopg2-binary
numpy